Delete entrybyplaintext package and move it to bibtexextractor
