+++ src/main/java/org/jabref/gui/bibtexextractor/BibtexExtractorViewModel.java
@@ -7,6 +7,7 @@ import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 
 import org.jabref.Globals;
+import org.jabref.logic.bibtexkeypattern.BibtexKeyGenerator;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.types.EntryType;
